Move option processing into mypy/options.py
